+++ Assets/Tests/LoaderUpgradeTests.cs
@@ -16,7 +16,7 @@ namespace Tests
         [TestCase(100, -0.95)]
         public void Loader_WorkCycleTime_Upgrade_Value_Is_Correct(int upgradeLevel, double expectedUpgradeValue) {
             var upgrade = new LoaderUpgrade();
-            var upgradeValue = upgrade.WorkCycleTime.GetUpgradeValue(upgradeLevel);
+            var upgradeValue = (double)upgrade.WorkCycleTime.GetUpgradeValue(upgradeLevel);
             Assert.AreEqual(expectedUpgradeValue, upgradeValue, delta: 0.001f);
         }
 
@@ -29,7 +29,7 @@ namespace Tests
         [TestCase(100, -0.95)]
         public void Loader_MovementSpeed_Upgrade_Value_Is_Correct(int upgradeLevel, double expectedUpgradeValue) {
             var upgrade = new LoaderUpgrade();
-            var upgradeValue = upgrade.MovementSpeed.GetUpgradeValue(upgradeLevel);
+            var upgradeValue = (double)upgrade.MovementSpeed.GetUpgradeValue(upgradeLevel);
             Assert.AreEqual(expectedUpgradeValue, upgradeValue, delta: 0.001f);
         }
 
@@ -42,7 +42,7 @@ namespace Tests
         [TestCase(100, "4444267362580")]
         public void Loader_Capacity_Upgrade_Value_Is_Correct(int upgradeLevel, string expectedUpgradeValue) {
             var upgrade = new LoaderUpgrade();
-            var upgradeValue = upgrade.Capacity.GetUpgradeValue(upgradeLevel);
+            var upgradeValue = (BigInteger)upgrade.Capacity.GetUpgradeValue(upgradeLevel);
             Assert.AreEqual(new BigInteger(expectedUpgradeValue.Replace(" ", string.Empty)), upgradeValue);
         }
 
@@ -54,7 +54,7 @@ namespace Tests
         [TestCase(400, "6")]
         public void Loader_NumberOfUnits_Upgrade_Value_Is_Correct(int upgradeLevel, string expectedUpgradeValue) {
             var upgrade = new LoaderUpgrade();
-            var upgradeValue = upgrade.NumberOfUnits.GetUpgradeValue(upgradeLevel);
+            var upgradeValue = (BigInteger)upgrade.NumberOfUnits.GetUpgradeValue(upgradeLevel);
             Assert.AreEqual(new BigInteger(expectedUpgradeValue.Replace(" ", string.Empty)), upgradeValue);
         }
     }
