+++ Assets/Tests/WarehouseUpgradeTests.cs
@@ -4,7 +4,6 @@ using NUnit.Framework;
 
 namespace Tests {
     public class WarehouseUpgradeTests {
-
         [TestCase(1, "20")]
         [TestCase(2, "22")]
         [TestCase(5, "28")]
@@ -27,7 +26,7 @@ namespace Tests {
         [TestCase(100, 2.05)]
         public void Warehouse_WorkCycleTime_Upgrade_Value_Is_Correct(int upgradeLevel, double expectedUpgradeValue) {
             var upgrade = new WarehouseUpgrade();
-            var upgradeValue = upgrade.WorkCycleTime.GetUpgradeValue(upgradeLevel);
+            var upgradeValue = (double) upgrade.WorkCycleTime.GetUpgradeValue(upgradeLevel);
             Assert.AreEqual(expectedUpgradeValue, upgradeValue, delta: 0.001f);
         }
 
@@ -40,7 +39,7 @@ namespace Tests {
         [TestCase(100, "11288701574364")]
         public void Warehouse_Capacity_Upgrade_Value_Is_Correct(int upgradeLevel, string expectedUpgradeValue) {
             var upgrade = new WarehouseUpgrade();
-            var upgradeValue = upgrade.Capacity.GetUpgradeValue(upgradeLevel);
+            var upgradeValue = (BigInteger) upgrade.Capacity.GetUpgradeValue(upgradeLevel);
             Assert.AreEqual(new BigInteger(expectedUpgradeValue.Replace(" ", string.Empty)), upgradeValue);
         }
     }
